backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
qdrant-client==1.7.3
sentence-transformers==2.3.1
scikit-learn==1.4.0
lightgbm==4.3.0
numpy==1.26.3
pandas==2.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
google-generativeai==0.3.2
celery==5.3.6
redis==5.0.1
